Replace Ints with BigIntegers for quantity figures

+++ apps/backend/src/main/kotlin/ca/uvictrades/trades/controller/admin/requests/AddStockToUserRequest.kt
@@ -1,6 +1,8 @@
 package ca.uvictrades.trades.controller.admin.requests
 
+import java.math.BigInteger
+
 data class AddStockToUserRequest(
 	val stock_id: String,
-	val quantity: Int,
+	val quantity: BigInteger,
 )
