+++ apps/backend/src/main/kotlin/ca/uvictrades/trades/matching/port/SellLimitOrder.kt
@@ -1,12 +1,13 @@
 package ca.uvictrades.trades.matching.port
 
 import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.Instant
 
 data class SellLimitOrder(
 	val id: String,
 	val stock: String,
-	val quantity: Int,
+	val quantity: BigInteger,
 	val pricePerUnit: BigDecimal,
 	val timestamp: Instant,
 	val trader: String
